@@ -8,8 +8,11 @@
     ],
     "type": "package",
     "minimum-stability": "dev",
-    "license": "Mit",
+    "license": "MIT",
     "autoload": {
+        "files": [
+            "src/Helpers/functions.php"
+        ],
         "psr-4": {
             "seregazhuk\\PinterestBot\\": "src"
         }
@@ -21,17 +24,18 @@
         }
     },
     "require": {
-        "php": ">=5.5.9",
-        "ext-curl": "*"
+        "php": ">=7.0",
+        "ext-curl": "*",
+        "ext-fileinfo": "*"
     },
     "require-dev": {
-        "codeclimate/php-test-reporter": "dev-master",
-        "mockery/mockery": "~0.9",
-        "phpunit/phpunit": "~4.0"
+        "codeclimate/php-test-reporter": "^0.4",
+        "mockery/mockery": "1.0",
+        "phpunit/phpunit": "~6.5"
     },
     "authors": [
         {
-            "name": "Serega Zhuk",
+            "name": "Sergey Zhuk",
             "email": "seregazhuk88@gmail.com",
             "role": "Developer"
         }
