@@ -2,19 +2,26 @@
 
 namespace seregazhuk\PinterestBot\Api\Providers;
 
+use LogicException;
 use seregazhuk\PinterestBot\Helpers\UrlBuilder;
+use seregazhuk\PinterestBot\Api\Forms\Registration;
+use seregazhuk\PinterestBot\Api\Providers\Core\Provider;
+use seregazhuk\PinterestBot\Api\Traits\ResolvesCurrentUser;
+use seregazhuk\PinterestBot\Api\Traits\SendsRegisterActions;
 
 class Auth extends Provider
 {
+    use SendsRegisterActions, ResolvesCurrentUser;
+
     /**
      * @var array
      */
     protected $loginRequiredFor = [
-        'logout'
+        'logout',
+        'convertToBusiness',
     ];
 
     const REGISTRATION_COMPLETE_EXPERIENCE_ID = '11:10105';
-    const ACCOUNT_TYPE_OTHER = 'other';
 
     /**
      * Login as pinner.
@@ -26,84 +33,95 @@ class Auth extends Provider
      */
     public function login($username, $password, $autoLogin = true)
     {
-        if ($this->request->isLoggedIn()) return true;
+        if ($this->isLoggedIn()) {
+            return true;
+        }
 
         $this->checkCredentials($username, $password);
 
         // Trying to load previously saved cookies from last login session for this username.
         // Then grab user profile info to check, if cookies are ok. If an empty response
         // was returned, then send login request.
-        if($autoLogin && $this->processAutoLogin($username)) {
+        if ($autoLogin && $this->processAutoLogin($username)) {
             return true;
         }
 
         return $this->processLogin($username, $password);
     }
 
-    /**
-     * If $removeCookies is set, cookie file will be removed
-     * from the file system.
-     *
-     * @param bool $removeCookies
-     */
-    public function logout($removeCookies = false)
+    public function logout()
     {
-        $this->request->logout($removeCookies);
+        $this->request->logout();
     }
 
     /**
      * Register a new user.
      *
-     * @param string $email
-     * @param string $password
-     * @param string $name
-     * @param string $country
-     * @param int $age
+     * @param string|Registration $email
+     * @param string|null $password
+     * @param string|null $name
+     * @param string $country @deprecated
+     * @param int $age @deprecated
      *
      * @return bool
      */
-    public function register($email, $password, $name, $country = 'GB', $age = 18)
+    public function register($email, $password = null, $name = null, $country = 'GB', $age = 18)
     {
-        $data = [
-            "age"        => $age,
-            "email"      => $email,
-            "password"   => $password,
-            "country"    => $country,
-            "first_name" => $name,
-            "container"  => 'home_page',
-        ];
+        $registrationForm = $this->getRegistrationForm($email, $password, $name, $country, $age);
 
-        return $this->makeRegisterCall($data);
+        return $this->makeRegisterCall($registrationForm);
     }
 
     /**
-     * Register a new business account.
+     * Register a new business account. At first we register a basic type account.
+     * Then convert it to a business one. This is done to receive a confirmation
+     * email after registration.
      *
-     * @param string $email
+     * @param string|Registration $registrationForm
      * @param string $password
-     * @param string $businessName
+     * @param string $name
      * @param string $website
      * @return bool|mixed
      */
-    public function registerBusiness($email, $password, $businessName, $website = '')
+    public function registerBusiness($registrationForm, $password = null, $name = null, $website = '')
     {
-        $data = [
-            "email"         => $email,
-            "password"      => $password,
-            "website_url"   => $website,
-            "business_name" => $businessName,
-            "account_type"  => self::ACCOUNT_TYPE_OTHER,
-        ];
+        $registration = $this->register($registrationForm, $password, $name);
 
-        return $this->makeRegisterCall($data);
+        if (!$registration) {
+            return false;
+        }
+
+        $website = ($registrationForm instanceof Registration) ?
+            $registrationForm->getSite() : $website;
+
+        return $this->convertToBusiness($name, $website);
     }
 
     /**
+     * Convert your account to a business one.
+     *
+     * @param string $businessName
+     * @param string $websiteUrl
      * @return bool
      */
-    public function isLoggedIn()
+    public function convertToBusiness($businessName, $websiteUrl = '')
     {
-        return $this->request->isLoggedIn();
+        $data = [
+            'business_name' => $businessName,
+            'website_url'   => $websiteUrl,
+            'account_type'  => 'other',
+        ];
+
+        return $this->post(UrlBuilder::RESOURCE_CONVERT_TO_BUSINESS, $data);
+    }
+
+    /**
+     * @param string $link
+     * @return array|bool
+     */
+    public function confirmEmail($link)
+    {
+        return $this->get($link);
     }
 
     /**
@@ -111,11 +129,12 @@ class Auth extends Provider
      *
      * @param string $username
      * @param string $password
+     * @throws LogicException
      */
     protected function checkCredentials($username, $password)
     {
         if (!$username || !$password) {
-            throw new \LogicException('You must set username and password to login.');
+            throw new LogicException('You must set username and password to login.');
         }
     }
 
@@ -124,24 +143,27 @@ class Auth extends Provider
      */
     protected function completeRegistration()
     {
-        $this->request->setTokenFromCookies();
-
-        return $this->execPostRequest(
-            ['placed_experience_id' => self::REGISTRATION_COMPLETE_EXPERIENCE_ID],
-            UrlBuilder::RESOURCE_REGISTRATION_COMPLETE
+        return $this->post(
+            UrlBuilder::RESOURCE_REGISTRATION_COMPLETE,
+            ['placed_experience_id' => self::REGISTRATION_COMPLETE_EXPERIENCE_ID]
         );
     }
 
     /**
-     * @param array $data
+     * @param Registration $registrationForm
      * @return bool|mixed
      */
-    protected function makeRegisterCall($data)
+    protected function makeRegisterCall(Registration $registrationForm)
     {
-        $this->visitMainPage();
-        $this->request->setTokenFromCookies();
+        if (!$this->sendEmailVerificationAction()) {
+            return false;
+        }
 
-        if (!$this->execPostRequest($data, UrlBuilder::RESOURCE_CREATE_REGISTER)) {
+        if (!$this->post(UrlBuilder::RESOURCE_CREATE_REGISTER, $registrationForm->toArray())) {
+            return false;
+        }
+
+        if (!$this->sendRegistrationActions()) {
             return false;
         }
 
@@ -155,17 +177,18 @@ class Auth extends Provider
      */
     protected function processLogin($username, $password)
     {
-        $this->request->clearToken();
-        $this->request->getHttpClient()->removeCookies();
+        $this->request->loadCookiesFor($username);
 
         $credentials = [
             'username_or_email' => $username,
             'password'          => $password,
         ];
 
-        $response = $this->execPostRequest($credentials, UrlBuilder::RESOURCE_LOGIN, true);
+        $this->post(UrlBuilder::RESOURCE_LOGIN, $credentials);
 
-        if (!$response->isOk()) return false;
+        if ($this->response->isEmpty()) {
+            return false;
+        }
 
         $this->request->login();
 
@@ -178,19 +201,41 @@ class Auth extends Provider
      */
     protected function processAutoLogin($username)
     {
-        return $this->request->autoLogin($username) && $this->getProfile();
-    }
-
-    public function visitMainPage()
-    {
-        $this->execGetRequest([], '');
+        return $this->request->autoLogin($username) && $this->resolveCurrentUserId();
     }
 
     /**
-     * @return array
+     * @param string $registrationForm
+     * @param string $password
+     * @param string $name
+     * @param string $country
+     * @param string $age
+     * @return Registration
      */
-    protected function getProfile()
+    protected function fillRegistrationForm($registrationForm, $password, $name, $country, $age)
     {
-        return $this->execGetRequest([], UrlBuilder::RESOURCE_GET_USER_SETTINGS);
+        return (new Registration($registrationForm, $password, $name))
+            ->setCountry($country)
+            ->setAge($age)
+            ->setGender('male');
+    }
+
+    /**
+     * @param string|Registration $email
+     * @param string $password
+     * @param string $name
+     * @param string $country
+     * @param string $age
+     * @return Registration
+     */
+    protected function getRegistrationForm($email, $password, $name, $country, $age)
+    {
+        if ($email instanceof Registration) {
+            return $email;
+        }
+
+        return $this->fillRegistrationForm(
+            $email, $password, $name, $country, $age
+        );
     }
 }
