@@ -2,27 +2,29 @@
 
 namespace seregazhuk\PinterestBot\Api\Providers;
 
-use Generator;
+use seregazhuk\PinterestBot\Helpers\Pagination;
 use seregazhuk\PinterestBot\Helpers\UrlBuilder;
 use seregazhuk\PinterestBot\Api\Traits\Searchable;
-use seregazhuk\PinterestBot\Api\Traits\Followable;
 use seregazhuk\PinterestBot\Api\Traits\CanBeDeleted;
+use seregazhuk\PinterestBot\Api\Traits\BoardInvites;
 use seregazhuk\PinterestBot\Api\Traits\SendsMessages;
-use seregazhuk\PinterestBot\Api\Contracts\PaginatedResponse;
+use seregazhuk\PinterestBot\Api\Traits\ResolvesCurrentUser;
+use seregazhuk\PinterestBot\Api\Providers\Core\FollowableProvider;
 
-class Boards extends Provider
+class Boards extends FollowableProvider
 {
-    use CanBeDeleted, Searchable, Followable, SendsMessages;
+    use CanBeDeleted, Searchable, SendsMessages, ResolvesCurrentUser, BoardInvites;
+
+    const BOARD_PRIVACY_PUBLIC = 'public';
+    const BOARD_PRIVACY_PRIVATE = 'secret';
 
     /**
      * @var array
      */
     protected $loginRequiredFor = [
-        'delete',
+        'my',
+        'forMe',
         'create',
-        'follow',
-        'unFollow',
-        'send',
     ];
 
     protected $searchScope  = 'boards';
@@ -33,17 +35,50 @@ class Boards extends Provider
     protected $unFollowUrl  = UrlBuilder::RESOURCE_UNFOLLOW_BOARD;
     protected $deleteUrl    = UrlBuilder::RESOURCE_DELETE_BOARD;
     protected $followersUrl = UrlBuilder::RESOURCE_BOARD_FOLLOWERS;
-    
+
+    protected $messageEntityName = 'board';
+
     /**
      * Get boards for user by username.
      *
      * @param string $username
      *
-     * @return array|bool
+     * @return array
      */
     public function forUser($username)
     {
-        return $this->execGetRequest(['username' => $username], UrlBuilder::RESOURCE_GET_BOARDS);
+        $options = [
+            'username' => $username,
+            'field_set_key' => 'detailed',
+        ];
+
+        $result = $this->get(UrlBuilder::RESOURCE_GET_BOARDS, $options);
+
+        return $result ?: [];
+    }
+
+    /**
+     * Get boards for current logged in user.
+     *
+     * @return array
+     */
+    public function forMe()
+    {
+        $currentUserName = $this->resolveCurrentUsername();
+
+        if (!$currentUserName) {
+            return [];
+        }
+
+        return $this->forUser($currentUserName);
+    }
+
+    /**
+     * @return array
+     */
+    public function my()
+    {
+        return $this->forMe();
     }
 
     /**
@@ -57,12 +92,22 @@ class Boards extends Provider
     public function info($username, $board)
     {
         $requestOptions = [
-            'username'      => $username,
-            'slug'          => $board,
+            'slug' => $this->formatBoardName($board),
+            'username' => $username,
             'field_set_key' => 'detailed',
         ];
 
-        return $this->execGetRequest($requestOptions, UrlBuilder::RESOURCE_GET_BOARD);
+        return $this->get(UrlBuilder::RESOURCE_GET_BOARD, $requestOptions);
+    }
+
+    /**
+     * @param string $board
+     * @return string
+     */
+    protected function formatBoardName($board)
+    {
+        $nameWithRemovedSpaces = str_replace(' ', '-', $board);
+        return function_exists('mb_strtolower') ? mb_strtolower($nameWithRemovedSpaces) : strtolower($nameWithRemovedSpaces);
     }
 
     /**
@@ -71,31 +116,18 @@ class Boards extends Provider
      * @param int $boardId
      * @param int $limit
      *
-     * @return Generator
+     * @return Pagination
      */
-    public function pins($boardId, $limit = 0)
+    public function pins($boardId, $limit = Pagination::DEFAULT_LIMIT)
     {
-        return $this->getPaginatedResponse(['boardId' => $boardId], $limit, 'getPinsFromBoard');
-    }
-
-    /**
-     * @param int   $boardId
-     * @param array $bookmarks
-     *
-     * @return PaginatedResponse
-     */
-    public function getPinsFromBoard($boardId, $bookmarks = [])
-    {
-        return $this->execGetRequestWithPagination(
-            ['board_id' => $boardId],
-            UrlBuilder::RESOURCE_GET_BOARD_FEED,
-            $bookmarks
+        return $this->paginate(
+            UrlBuilder::RESOURCE_GET_BOARD_FEED, ['board_id' => $boardId], $limit
         );
     }
 
     /**
      * Update board info. Gets boardId and an associative array as params. Available keys of the array are:
-     * 'category', 'description', 'privacy'.
+     * 'name', 'category', 'description', 'privacy'.
      *
      * - 'privacy' can be 'public' or 'secret'. 'public' by default.
      * - 'category' is 'other' by default.
@@ -106,6 +138,10 @@ class Boards extends Provider
      */
     public function update($boardId, $attributes)
     {
+        if (isset($attributes['name'])) {
+            $attributes['name'] = $this->formatBoardName($attributes['name']);
+        }
+
         $requestOptions = array_merge(
             [
                 'board_id' => $boardId,
@@ -113,7 +149,7 @@ class Boards extends Provider
             ], $attributes
         );
 
-        return $this->execPostRequest($requestOptions, UrlBuilder::RESOURCE_UPDATE_BOARD);
+        return $this->post(UrlBuilder::RESOURCE_UPDATE_BOARD, $requestOptions);
     }
 
     /**
@@ -121,19 +157,32 @@ class Boards extends Provider
      *
      * @param string $name
      * @param string $description
-     * @param string $privacy     Can be 'public' or 'secret'. 'public' by default.
+     * @param string $privacy Can be 'public' or 'secret'. 'public' by default.
      *
      * @return bool
      */
-    public function create($name, $description, $privacy = 'public')
+    public function create($name, $description, $privacy = self::BOARD_PRIVACY_PUBLIC)
     {
         $requestOptions = [
-            'name'        => $name,
+            'name' => $name,
             'description' => $description,
-            'privacy'     => $privacy,
+            'privacy' => $privacy,
         ];
 
-        return $this->execPostRequest($requestOptions, UrlBuilder::RESOURCE_CREATE_BOARD);
+        return $this->post(UrlBuilder::RESOURCE_CREATE_BOARD, $requestOptions);
+    }
+
+    /**
+     * Create a new board.
+     *
+     * @param string $name
+     * @param string $description
+     *
+     * @return bool
+     */
+    public function createPrivate($name, $description)
+    {
+        return $this->create($name, $description, self::BOARD_PRIVACY_PRIVATE);
     }
 
     /**
@@ -144,22 +193,23 @@ class Boards extends Provider
      */
     public function titleSuggestionsFor($pinId)
     {
-        return $this->execGetRequest(['pin_id' => $pinId], UrlBuilder::RESOURCE_TITLE_SUGGESTIONS);
+        return $this->get(UrlBuilder::RESOURCE_TITLE_SUGGESTIONS, ['pin_id' => $pinId]);
     }
 
     /**
-     * Send board with message or by email.
-     *
-     * @param int $boardId
-     * @param string $text
-     * @param array|string $userIds
-     * @param array|string $emails
-     * @return bool
+     * @param string $boardId
+     * @return array|bool
      */
-    public function send($boardId, $text, $userIds, $emails)
+    public function leave($boardId)
     {
-        $messageData = $this->buildMessageData($text, $boardId);
+        $requestOptions = [
+            'ban' => false,
+            'board_id' => (string)$boardId,
+            'field_set_key' => 'boardEdit',
+            'invited_user_id' => (string)$this->resolveCurrentUserId(),
+        ];
+        print_r($requestOptions);
 
-        return $this->callSendMessage($userIds, $emails, $messageData);
+        return $this->post(UrlBuilder::RESOURCE_LEAVE_BOARD, $requestOptions);
     }
 }
