@@ -2,7 +2,9 @@
 
 namespace seregazhuk\PinterestBot\Api\Providers;
 
+use seregazhuk\PinterestBot\Helpers\Pagination;
 use seregazhuk\PinterestBot\Helpers\UrlBuilder;
+use seregazhuk\PinterestBot\Api\Providers\Core\Provider;
 
 class Comments extends Provider
 {
@@ -17,19 +19,20 @@ class Comments extends Provider
     /**
      * Write a comment for a pin with current id.
      *
-     * @param int    $pinId
-     * @param string $text  Comment
+     * @param int $pinId
+     * @param string $text Comment
      *
-     * @return array|bool
+     * @return array
      */
     public function create($pinId, $text)
     {
         $requestOptions = [
-            'pin_id' => $pinId,
-            'text'   => $text,
+            'objectId' => $this->getAggregatedPinId($pinId),
+            'pinId'    => $pinId,
+            'text'     => $text,
         ];
 
-        return $this->execPostRequest($requestOptions, UrlBuilder::RESOURCE_COMMENT_PIN);
+        return $this->post(UrlBuilder::RESOURCE_COMMENT_PIN, $requestOptions, true);
     }
 
     /**
@@ -42,12 +45,39 @@ class Comments extends Provider
      */
     public function delete($pinId, $commentId)
     {
-        $requestOptions = [
-            'pin_id'     => $pinId,
-            'comment_id' => $commentId,
-        ];
+        $requestOptions = ['commentId' => $commentId];
 
-        return $this->execPostRequest($requestOptions, UrlBuilder::RESOURCE_COMMENT_DELETE_PIN);
+        return $this->post(UrlBuilder::RESOURCE_COMMENT_DELETE_PIN, $requestOptions);
     }
 
+
+    /**
+     * @param string $pinId
+     * @param int $limit
+     * @return Pagination
+     */
+    public function getList($pinId, $limit = Pagination::DEFAULT_LIMIT)
+    {
+        return $this->paginate(
+            'resource/AggregatedCommentResource/get/',
+            ['bookmarks' => '', 'objectId' => $this->getAggregatedPinId($pinId), 'page_size' => 2],
+            $limit
+        );
+    }
+
+    /**
+     * @param string $pinId
+     * @return null|string
+     */
+    protected function getAggregatedPinId($pinId)
+    {
+        $requestOptions = [
+            'id'            => $pinId,
+            'field_set_key' => 'detailed',
+        ];
+
+        $pinInfo = $this->get(UrlBuilder::RESOURCE_PIN_INFO, $requestOptions);
+
+        return isset($pinInfo['aggregated_pin_data']['id']) ? $pinInfo['aggregated_pin_data']['id'] : null;
+    }
 }
