@@ -2,23 +2,14 @@
 
 namespace seregazhuk\PinterestBot\Api\Providers;
 
-use Generator;
-use seregazhuk\PinterestBot\Api\Traits\HasFeed;
+use seregazhuk\PinterestBot\Helpers\Pagination;
 use seregazhuk\PinterestBot\Helpers\UrlBuilder;
-use seregazhuk\PinterestBot\Api\Traits\Followable;
 use seregazhuk\PinterestBot\Api\Traits\HasRelatedTopics;
+use seregazhuk\PinterestBot\Api\Providers\Core\FollowableProvider;
 
-class Topics extends Provider
+class Topics extends FollowableProvider
 {
-    use Followable, HasRelatedTopics, HasFeed;
-
-    /**
-     * @var array
-     */
-    protected $loginRequiredFor = [
-        'follow',
-        'unFollow',
-    ];
+    use HasRelatedTopics;
 
     protected $followUrl   = UrlBuilder::RESOURCE_FOLLOW_INTEREST;
     protected $unFollowUrl = UrlBuilder::RESOURCE_UNFOLLOW_INTEREST;
@@ -33,7 +24,7 @@ class Topics extends Provider
      */
     public function info($topic)
     {
-        return $this->execGetRequest(["interest" => $topic], UrlBuilder::RESOURCE_GET_TOPIC);
+        return $this->get(UrlBuilder::RESOURCE_GET_TOPIC, ['interest' => $topic]);
     }
 
     /**
@@ -41,15 +32,32 @@ class Topics extends Provider
      *
      * @param string $interest
      * @param int $limit
-     * @return Generator
+     * @return Pagination
      */
-    public function pins($interest, $limit = 0)
+    public function pins($interest, $limit = Pagination::DEFAULT_LIMIT)
     {
         $data = [
             'interest'  => $interest,
             'pins_only' => false,
         ];
 
-        return $this->getFeed($data, UrlBuilder::RESOURCE_GET_TOPIC_FEED, $limit);
+        return $this->paginate(UrlBuilder::RESOURCE_GET_TOPIC_FEED, $data, $limit);
+    }
+
+    /**
+     * Returns an array of trending topics from http://pinterest.com/discover page. Then
+     * you can use an id of each topic to get trending pins for this topic with
+     * $bot->pins->explore() method.
+     *
+     * @return array
+     */
+    public function explore()
+    {
+        $data = [
+            'aux_fields' => [],
+            'offset'     => 180,
+        ];
+
+        return $this->get(UrlBuilder::RESOURCE_EXPLORE_SECTIONS, $data);
     }
 }
