@@ -3,6 +3,7 @@
 namespace seregazhuk\PinterestBot\Api;
 
 use seregazhuk\PinterestBot\Api\Contracts\PaginatedResponse;
+use function seregazhuk\PinterestBot\get_array_data;
 
 class Response implements PaginatedResponse
 {
@@ -12,7 +13,7 @@ class Response implements PaginatedResponse
     protected $data = [];
 
     /**
-     * @var array|null
+     * @var Error
      */
     protected $lastError;
 
@@ -21,20 +22,44 @@ class Response implements PaginatedResponse
      */
     protected $clientInfo;
 
+    /**
+     * @param mixed $data
+     */
+    public function __construct($data = null)
+    {
+        $this->lastError = new Error();
+
+        if ($data) {
+            $this->fill($data);
+        }
+    }
+
+    /**
+     * @param mixed $data
+     * @return $this
+     */
     public function fill($data)
     {
         $this->data = $data;
 
-        $this->lastError = $this->getValueByKey('resource_response.error', $this->data);
+        return $this
+            ->fillError()
+            ->fillClientInfo();
+    }
 
-        $this->clientInfo = $this->getValueByKey('client_context', $this->data);
+    /**
+     * @param string $json
+     * @return static
+     */
+    public function fillFromJson($json)
+    {
+        return $this->fill(json_decode($json, true));
     }
 
     /**
      * Check if specified data exists in response.
      *
-     * @param null  $key
-     *
+     * @param null $key
      * @return array|bool
      */
     public function getResponseData($key = null)
@@ -48,12 +73,12 @@ class Response implements PaginatedResponse
 
     /**
      * @param string $key
-     * @param null $default
+     * @param null|mixed $default
      * @return mixed
      */
     public function getData($key = '', $default = null)
     {
-        return $this->getValueByKey($key, $this->data, $default);
+        return get_array_data($key, $this->data, $default);
     }
 
     /**
@@ -62,47 +87,27 @@ class Response implements PaginatedResponse
      */
     public function hasData($key = '')
     {
-        return !is_null($this->getValueByKey($key, $this->data));
+        return get_array_data($key, $this->data) !== null;
     }
 
     /**
      * Parse data from Pinterest Api response.
      * Data is stored in ['resource_response']['data'] array.
      *
-     * @param string $key
-     *
+     * @param string|null $key
      * @return bool|array
      */
-    protected function parseResponseData($key)
+    protected function parseResponseData($key = null)
     {
-        $responseData = $this->getValueByKey('resource_response.data', $this->data);
-        if(!$responseData) return false;
+        $responseData = get_array_data('resource_response.data', $this->data);
 
-        return $key ?
-            $this->getValueByKey($key, $responseData) :
-            $responseData;
-    }
-
-    /**
-     * @param string $key
-     * @param array $data
-     * @param mixed $default
-     * @return array|bool|mixed
-     */
-    protected function getValueByKey($key = '', $data, $default = null)
-    {
-        if(empty($key)) return $data;
-
-        $indexes = explode('.', $key);
-        $value = $data;
-
-        foreach ($indexes as $index) {
-            if(!isset($value[$index])) return $default;
-
-            $value = $value[$index];
+        if (!$responseData) {
+            return false;
         }
 
-        return $value;
+        return $key ?
+            get_array_data($key, $responseData) :
+            $responseData;
     }
 
     /**
@@ -120,26 +125,17 @@ class Response implements PaginatedResponse
      */
     public function isOk()
     {
-        return !$this->hasErrors() && !$this->isEmpty();
+        return !$this->hasErrors();
     }
 
     /**
-     * @return bool
-     */
-    public function hasResponseData()
-    {
-        return (bool)$this->getValueByKey('resource_response.data', $this->data);
-    }
-
-    /**
-     * Check for error info in api response and save
-     * it.
+     * Check for error info in api response and save it.
      *
      * @return bool
      */
     public function hasErrors()
     {
-        return !is_null($this->lastError);
+        return !$this->lastError->isEmpty();
     }
 
     /**
@@ -149,8 +145,30 @@ class Response implements PaginatedResponse
      */
     public function getBookmarks()
     {
-        $bookmarks = $this->getValueByKey('resource.options.bookmarks', $this->data,  []);
-        return empty($bookmarks) ? [] : [$bookmarks[0]];
+        $bookmarks = $this->getRawBookmarksData();
+
+        if (empty($bookmarks)) {
+            return [];
+        }
+
+        if ($bookmarks[0] === '-end-') {
+            return [];
+        }
+
+        return $bookmarks;
+    }
+
+    /**
+     * @return bool
+     */
+    public function hasBookmarks()
+    {
+        return (bool) $this->getBookmarks();
+    }
+
+    protected function getRawBookmarksData()
+    {
+        return $this->getData('resource.options.bookmarks', []);
     }
 
     /**
@@ -161,24 +179,26 @@ class Response implements PaginatedResponse
      */
     public function getPaginationData()
     {
-        if ($this->isEmpty() && $this->hasErrors()) {
+        if ($this->hasErrors()) {
             return [];
         }
 
-        $bookmarks = $this->getBookmarks();
         if ($data = $this->getResponseData()) {
-            return ['data' => $data, 'bookmarks' => $bookmarks];
+            return [
+                'data'      => $data,
+                'bookmarks' => $this->getBookmarks(),
+            ];
         }
 
         return [];
     }
 
     /**
-     * @return array
+     * @return string|null
      */
-    public function getLastError()
+    public function getLastErrorText()
     {
-        return $this->lastError;
+        return $this->lastError->getText();
     }
 
     /**
@@ -187,5 +207,34 @@ class Response implements PaginatedResponse
     public function getClientInfo()
     {
         return $this->clientInfo;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getRawData()
+    {
+        return $this->data;
+    }
+
+    public function clear()
+    {
+        return $this->fill([]);
+    }
+
+    protected function fillError()
+    {
+        $errorData = get_array_data('resource_response.error', $this->data);
+
+        $this->lastError = new Error($errorData);
+
+        return $this;
+    }
+
+    protected function fillClientInfo()
+    {
+        $this->clientInfo = get_array_data('client_context', $this->data);
+
+        return $this;
     }
 }
